+++ docs/source/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+Added
+~~~~~
+* Added ``Graph.adjacency_matrix()`` method.
+* `Issue #29 <https://github.com/pmacg/py-sgtl/issues/29>`_ - add spectrum methods for adjacency and laplacian matrices.
+
 0.4 - 2021-12-14
 ----------------